We've updated our privacy policy
We are committed to protect the privacy of Orientations. Because of
recent changes in Cart data access law, we have updated our attribute
policy, which is effective as of <local.get('now', 'GST')>.

The updated Privacy Policy reflects the increased protection and
adhesion to CS101 course material taught in Java, where more privater ->
more betterer.